chore: More, more and more

+++ server/filesystem/stat_linux.go
@@ -14,11 +14,11 @@ import (
 func (s *Stat) CTime() time.Time {
 	if st, ok := s.Sys().(*unix.Stat_t); ok {
 		// Do not remove these "redundant" type-casts, they are required for 32-bit builds to work.
-		return time.Unix(int64(st.Ctim.Sec), int64(st.Ctim.Nsec))
+		return time.Unix(st.Ctim.Sec, st.Ctim.Nsec)
 	}
 	if st, ok := s.Sys().(*syscall.Stat_t); ok {
 		// Do not remove these "redundant" type-casts, they are required for 32-bit builds to work.
-		return time.Unix(int64(st.Ctim.Sec), int64(st.Ctim.Nsec))
+		return time.Unix(st.Ctim.Sec, st.Ctim.Nsec)
 	}
 	return time.Time{}
 }
